Prevent Cleaver from cleaving peaceful bystanders (from xNetHack)
Original commit message from copperwater/xnethack@70b695ec13

A fairly common criticism of Cleaver's new behavior of hitting 3
monsters at once is that it can accidentally hit peacefuls without
warning the player, which can lead to all sorts of problems like
angering Minetown or the quest friendlies. So, this commit adds a check
function should_cleave that determines if there are peacefuls the player
can see that would be hit by Cleaver's arc while not being the main
target of it. If it finds any, it prompts the player to confirm
attacking the peacefuls; if they say yes, the arc attack proceeds and
will hit the peacefuls, but if they say no, they just hit the main
target normally.

Originally I had this as it just always treating this as if the player
said no, but then realized there are probably cases where the player
would determine it's okay to catch a peaceful in the arc for various
reasons, so they should have that option.

This doesn't go as generously as some other Cleaver proposals, such as
that any peacefuls should just be magically ignored so the arc only hits
two or one of the possible targets as needed.